+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+### Fixed
+- **CUMULUS-802** - Stop polling for new work if a SIGTERM is received
+
 ## [v1.2.4]
 
 ### Fixed